third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_character_17.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'character'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} character`
  );
}

function makeSample(offset){
  const kSample = '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right character
testIt('ltr', 0, 'right', '^\u202B|;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 1, 'right', '\u202B^;|car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 2, 'right', '\u202B;^c|ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 3, 'right', '\u202B;c^a|r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 4, 'right', '\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 5, 'right', '\u202B;car^ |\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 6, 'right', '\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 7, 'right', '\u202B;car \u05D3^;|\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 8, 'right', '\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 9, 'right', '\u202B;car \u05D3;\u05D4^;|\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 10, 'right', '\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 11, 'right', '\u202B;car \u05D3;\u05D4;\u05D5^;| \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 12, 'right', '\u202B;car \u05D3;\u05D4;\u05D5;^ |\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 13, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 14, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;|\u05D1;\u05D2;.\u202C;');
testIt('ltr', 15, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;.\u202C;');
testIt('ltr', 16, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;|\u05D2;.\u202C;');
testIt('ltr', 17, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;.\u202C;');
testIt('ltr', 18, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|.\u202C;');
testIt('ltr', 19, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^.|\u202C;');
testIt('ltr', 20, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.^\u202C|;');
testIt('ltr', 21, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C^;|');

// ltr left character
testIt('ltr', 0, 'left', '|\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 1, 'left', '|\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 2, 'left', '\u202B|;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 3, 'left', '\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 4, 'left', '\u202B;c|a^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 5, 'left', '\u202B;ca|r^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 6, 'left', '\u202B;car| ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 7, 'left', '\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 8, 'left', '\u202B;car \u05D3|;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 9, 'left', '\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 10, 'left', '\u202B;car \u05D3;\u05D4|;^\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 11, 'left', '\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 12, 'left', '\u202B;car \u05D3;\u05D4;\u05D5|;^ \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 13, 'left', '\u202B;car \u05D3;\u05D4;\u05D5;| ^\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 14, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 15, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0|;^\u05D1;\u05D2;.\u202C;');
testIt('ltr', 16, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;.\u202C;');
testIt('ltr', 17, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1|;^\u05D2;.\u202C;');
testIt('ltr', 18, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;.\u202C;');
testIt('ltr', 19, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2|;^.\u202C;');
testIt('ltr', 20, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|.^\u202C;');
testIt('ltr', 21, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.|\u202C^;');

// ltr forward character
testIt('ltr', 0, 'forward', '^\u202B|;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 1, 'forward', '\u202B^;|car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 2, 'forward', '\u202B;^c|ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 3, 'forward', '\u202B;c^a|r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 4, 'forward', '\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 5, 'forward', '\u202B;car^ |\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 6, 'forward', '\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 7, 'forward', '\u202B;car \u05D3^;|\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 8, 'forward', '\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 9, 'forward', '\u202B;car \u05D3;\u05D4^;|\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 10, 'forward', '\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 11, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5^;| \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 12, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5;^ |\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 13, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 14, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;|\u05D1;\u05D2;.\u202C;');
testIt('ltr', 15, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;.\u202C;');
testIt('ltr', 16, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;|\u05D2;.\u202C;');
testIt('ltr', 17, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;.\u202C;');
testIt('ltr', 18, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|.\u202C;');
testIt('ltr', 19, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^.|\u202C;');
testIt('ltr', 20, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.^\u202C|;');
testIt('ltr', 21, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C^;|');

// ltr backward character
testIt('ltr', 0, 'backward', '|\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 1, 'backward', '|\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 2, 'backward', '\u202B|;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 3, 'backward', '\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 4, 'backward', '\u202B;c|a^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 5, 'backward', '\u202B;ca|r^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 6, 'backward', '\u202B;car| ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 7, 'backward', '\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 8, 'backward', '\u202B;car \u05D3|;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 9, 'backward', '\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 10, 'backward', '\u202B;car \u05D3;\u05D4|;^\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 11, 'backward', '\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 12, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5|;^ \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 13, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5;| ^\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 14, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;.\u202C;');
testIt('ltr', 15, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0|;^\u05D1;\u05D2;.\u202C;');
testIt('ltr', 16, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;.\u202C;');
testIt('ltr', 17, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1|;^\u05D2;.\u202C;');
testIt('ltr', 18, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;.\u202C;');
testIt('ltr', 19, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2|;^.\u202C;');
testIt('ltr', 20, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|.^\u202C;');
testIt('ltr', 21, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.|\u202C^;');

// rtl right character
testIt('rtl', 0, 'right', '^\u202B|;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 1, 'right', '\u202B^;|car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 2, 'right', '\u202B;^c|ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 3, 'right', '\u202B;c^a|r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 4, 'right', '\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 5, 'right', '\u202B;car^ |\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 6, 'right', '\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 7, 'right', '\u202B;car \u05D3^;|\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 8, 'right', '\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 9, 'right', '\u202B;car \u05D3;\u05D4^;|\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 10, 'right', '\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 11, 'right', '\u202B;car \u05D3;\u05D4;\u05D5^;| \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 12, 'right', '\u202B;car \u05D3;\u05D4;\u05D5;^ |\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 13, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 14, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;|\u05D1;\u05D2;.\u202C;');
testIt('rtl', 15, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;.\u202C;');
testIt('rtl', 16, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;|\u05D2;.\u202C;');
testIt('rtl', 17, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;.\u202C;');
testIt('rtl', 18, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|.\u202C;');
testIt('rtl', 19, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^.|\u202C;');
testIt('rtl', 20, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.^\u202C|;');
testIt('rtl', 21, 'right', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C^;|');

// rtl left character
testIt('rtl', 0, 'left', '|\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 1, 'left', '|\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 2, 'left', '\u202B|;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 3, 'left', '\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 4, 'left', '\u202B;c|a^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 5, 'left', '\u202B;ca|r^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 6, 'left', '\u202B;car| ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 7, 'left', '\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 8, 'left', '\u202B;car \u05D3|;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 9, 'left', '\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 10, 'left', '\u202B;car \u05D3;\u05D4|;^\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 11, 'left', '\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 12, 'left', '\u202B;car \u05D3;\u05D4;\u05D5|;^ \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 13, 'left', '\u202B;car \u05D3;\u05D4;\u05D5;| ^\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 14, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 15, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0|;^\u05D1;\u05D2;.\u202C;');
testIt('rtl', 16, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;.\u202C;');
testIt('rtl', 17, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1|;^\u05D2;.\u202C;');
testIt('rtl', 18, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;.\u202C;');
testIt('rtl', 19, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2|;^.\u202C;');
testIt('rtl', 20, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|.^\u202C;');
testIt('rtl', 21, 'left', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.|\u202C^;');

// rtl forward character
testIt('rtl', 0, 'forward', '^\u202B|;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 1, 'forward', '\u202B^;|car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 2, 'forward', '\u202B;^c|ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 3, 'forward', '\u202B;c^a|r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 4, 'forward', '\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 5, 'forward', '\u202B;car^ |\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 6, 'forward', '\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 7, 'forward', '\u202B;car \u05D3^;|\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 8, 'forward', '\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 9, 'forward', '\u202B;car \u05D3;\u05D4^;|\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 10, 'forward', '\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 11, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5^;| \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 12, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5;^ |\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 13, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 14, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;|\u05D1;\u05D2;.\u202C;');
testIt('rtl', 15, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;.\u202C;');
testIt('rtl', 16, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;|\u05D2;.\u202C;');
testIt('rtl', 17, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;.\u202C;');
testIt('rtl', 18, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|.\u202C;');
testIt('rtl', 19, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^.|\u202C;');
testIt('rtl', 20, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.^\u202C|;');
testIt('rtl', 21, 'forward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C^;|');

// rtl backward character
testIt('rtl', 0, 'backward', '|\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 1, 'backward', '|\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 2, 'backward', '\u202B|;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 3, 'backward', '\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 4, 'backward', '\u202B;c|a^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 5, 'backward', '\u202B;ca|r^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 6, 'backward', '\u202B;car| ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 7, 'backward', '\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 8, 'backward', '\u202B;car \u05D3|;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 9, 'backward', '\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 10, 'backward', '\u202B;car \u05D3;\u05D4|;^\u05D5; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 11, 'backward', '\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 12, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5|;^ \u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 13, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5;| ^\u05D0;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 14, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;.\u202C;');
testIt('rtl', 15, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0|;^\u05D1;\u05D2;.\u202C;');
testIt('rtl', 16, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;.\u202C;');
testIt('rtl', 17, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1|;^\u05D2;.\u202C;');
testIt('rtl', 18, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;.\u202C;');
testIt('rtl', 19, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2|;^.\u202C;');
testIt('rtl', 20, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|.^\u202C;');
testIt('rtl', 21, 'backward', '\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;.|\u202C^;');
</script>
